+++ ikea-zigbee-drivers/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.3.0] - 2024-??-??
+- Add driver for Gewiss 2-channel Contact Interface 230V (GWA1502)
+
+### Added
+- Add driver for Tradfri Signal Repeater (E1746)
+- E2013: Add command to manually set contact status - `@rgr`
+
 ## [5.2.0] - 2024-11-09
 
 ### Added